practice multiplying two Scanner inputs

+++ J01_JavaBasics/C07_product.java
@@ -3,9 +3,9 @@ import java.util.*;
 public class C07_product {
     public static void main(String args[]) {
         Scanner sc = new Scanner(System.in);
-        int a = sc.nextInt();
-        int b = sc.nextInt();
-        int product = a * b;
+        float a = sc.nextFloat();
+        float b = sc.nextFloat();
+        float product = a * b;
         System.out.println(product);
     }
 }
